requirement.txt
asgiref==3.5.0
Django==4.0.3
django-ckeditor==6.2.0
django-crispy-forms==1.14.0
django-easy-pdf==0.1.1
django-js-asset==2.0.0
djangorestframework==3.13.1
future==0.18.2
html5lib==1.1
lxml==4.8.0
Pillow==9.0.1
PyPDF3==1.0.6
python-bidi==0.4.2
pytz==2021.3
reportlab==3.6.9
tqdm==4.63.1
xhtml2pdf==0.2.6
